Restructure script files
