+++ .gitignore
@@ -2,3 +2,5 @@ __pycache__
 env
 .vscode
 .DS_Store
+*.egg-info
+*.zip
